make setup a first class module
